requirements.txt
pyplis
# packages in environment at C:\Users\tw9616\Anaconda3\envs\py38:
#
# Name                    Version                   Build  Channel
bcrypt                    3.1.7            py38he774522_0  
blas                      1.0                         mkl  
ca-certificates           2020.1.1                      0  
certifi                   2020.4.5.1               py38_0  
cffi                      1.14.0           py38h7a1dbc1_0  
cryptography              2.9.2            py38h7a1dbc1_0  
cycler                    0.10.0                   py38_0  
freetype                  2.10.2               hd328e21_0    conda-forge
hdf5                      1.10.4               h7ebc959_0  
icc_rt                    2019.0.0             h0cc432a_1  
icu                       58.2                 ha925a31_3  
idna                      2.9                        py_1  
intel-openmp              2020.1                      216  
jpeg                      9c                hfa6e2cd_1001    conda-forge
kiwisolver                1.2.0            py38h74a9793_0  
libblas                   3.8.0           8_h8933c1f_netlib    conda-forge
libcblas                  3.8.0           8_h8933c1f_netlib    conda-forge
libclang                  9.0.1           default_hf44288c_0    conda-forge
liblapack                 3.8.0           8_h8933c1f_netlib    conda-forge
liblapacke                3.8.0           8_h8933c1f_netlib    conda-forge
libopencv                 4.0.1                hbb9e17c_0  
libpng                    1.6.37               hfe6a214_1    conda-forge
libtiff                   4.1.0                h885aae3_6    conda-forge
libwebp                   1.0.2                hfa6e2cd_5    conda-forge
lz4-c                     1.9.2                h62dcd97_1    conda-forge
m2w64-gcc-libgfortran     5.3.0                         6  
m2w64-gcc-libs            5.3.0                         7  
m2w64-gcc-libs-core       5.3.0                         7  
m2w64-gmp                 6.1.0                         2  
m2w64-libwinpthread-git   5.0.0.4634.697f757               2  
matplotlib                3.1.3                    py38_0  
matplotlib-base           3.1.3            py38h64f37c6_0  
mkl                       2020.1                      216  
mkl-service               2.3.0            py38hb782905_0  
mkl_fft                   1.0.15           py38h14836fe_0  
mkl_random                1.1.0            py38hf9181ef_0  
msys2-conda-epoch         20160418                      1  
numpy                     1.18.1           py38h93ca92e_0  
numpy-base                1.18.1           py38hc3f5095_1  
opencv                    4.0.1            py38h2a7c758_0  
openssl                   1.1.1g               he774522_0  
paramiko                  2.7.1                      py_0  
pillow                    7.1.2                    pypi_0    pypi
pip                       20.0.2                   py38_3  
py-opencv                 4.0.1            py38he44ac1e_0  
pycparser                 2.20                       py_0  
pynacl                    1.3.0            py38h62dcd97_0  
pyparsing                 2.4.7                      py_0  
pyqt                      5.9.2            py38ha925a31_4  
python                    3.8.2               he1778fa_13  
python-dateutil           2.8.1                      py_0  
python_abi                3.8                      1_cp38    conda-forge
qt                        5.9.7            vc14h73c81de_0  
setuptools                46.2.0                   py38_0  
sip                       4.19.13          py38ha925a31_0  
six                       1.14.0                   py38_0  
sqlite                    3.31.1               h2a8f88b_1  
tornado                   6.0.4            py38he774522_1  
ttkthemes                 3.0.0                    pypi_0    pypi
vc                        14.1                 h0510ff6_4  
vs2015_runtime            14.16.27012          hf0eaf9b_1  
wheel                     0.34.2                   py38_0  
wincertstore              0.2                      py38_0  
xz                        5.2.5                h2fa13f4_0    conda-forge
zlib                      1.2.11               h62dcd97_4  
zstd                      1.4.4                h9f78265_3    conda-forge
